requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
google-cloud-datastore==1.15.0
google-cloud-storage==1.31.2
google-cloud-ndb==1.6.0
google-cloud-vision==2.0.0
openai==0.23.0
gunicorn==20.1.0